[host][x86] mv assign_value to host, test=develop
