+++ lite/kernels/x86/assign_value_compute.cc
@@ -12,62 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "lite/kernels/x86/assign_value_compute.h"
-#include <string>
-#include <vector>
-#include "lite/core/op_registry.h"
-#include "lite/core/tensor.h"
-#include "lite/core/type_system.h"
-#include "lite/core/types.h"
-
-namespace paddle {
-namespace lite {
-namespace kernels {
-namespace x86 {
-
-template <class T>
-void TensorFromVector(const std::vector<T>& src, lite::Tensor* dst) {
-  auto* src_ptr = static_cast<const void*>(src.data());
-  auto* dst_ptr = static_cast<void*>(dst->mutable_data<T>());
-  auto size = src.size() * sizeof(T);
-  // std::memcpy(dst_ptr, src_ptr, size);
-  lite::TargetWrapperHost::MemcpySync(
-      dst_ptr, src_ptr, size, IoDirection::HtoH);
-}
-
-void AssignValueCompute::Run() {
-  auto& param = Param<operators::AssignValueParam>();
-  int dtype = param.dtype;
-  std::vector<float> fp32_values = param.fp32_values;
-  std::vector<int> int32_values = param.int32_values;
-  std::vector<int64_t> int64_values = param.int64_values;
-  std::vector<int> bool_values = param.bool_values;
-  auto* out = param.Out;
-
-  if (dtype == static_cast<int>(lite::core::FluidType::INT32)) {
-    TensorFromVector(int32_values, out);
-  } else if (dtype == static_cast<int>(lite::core::FluidType::FP32)) {
-    TensorFromVector(fp32_values, out);
-  } else if (dtype == static_cast<int>(lite::core::FluidType::INT64)) {
-    TensorFromVector(int64_values, out);
-  } else if (dtype == static_cast<int>(lite::core::FluidType::BOOL)) {
-    TensorFromVector(bool_values, out);
-  } else {
-    LOG(FATAL) << "Unsupported dtype for assign_value_op:" << dtype;
-  }
-  return;
-}
-
-}  // namespace x86
-}  // namespace kernels
-}  // namespace lite
-}  // namespace paddle
+#include "lite/kernels/host/assign_value_compute.h"
 
 REGISTER_LITE_KERNEL(assign_value,
                      kX86,
                      kAny,
                      kNCHW,
-                     paddle::lite::kernels::x86::AssignValueCompute,
+                     paddle::lite::kernels::host::AssignValueCompute,
                      def)
     .BindOutput("Out", {LiteType::GetTensorTy(TARGET(kX86), PRECISION(kAny))})
     .Finalize();
