+++ lite/tests/kernels/assign_value_compute_test.cc
@@ -143,10 +143,8 @@ class AssignValueComputeTester : public arena::TestCase {
 
 TEST(AssignValue, precision) {
   Place place;
-#ifdef LITE_WITH_ARM
-  place = TARGET(kARM);
-#elif LITE_WITH_X86
-  place = TARGET(kX86);
+#ifdef defined(LITE_WITH_ARM) || defined(LITE_WITH_X86)
+  place = TARGET(kHost);
 #else
   return;
 #endif
